Fixed GetPosts page/limit to be Int64 like in Lemmy backend

+++ Sources/LemmyKit/Types/GetPosts.swift
@@ -17,12 +17,16 @@ public struct GetPosts: LemmyApiEndpoint {
         /// Specifies how the posts in response should be ordered.
         public let sort: SortType?
 
-        /// Number of page to fetch. First page is number 1.
-        public let page: Int?
+        /// Pagination. The page number to fetch the results. This allows to fetch large data sets one page at a time.
+        /// The first page is number 1.
+        public let page: Int64?
 
+        /// Pagination. Specifies the maximum number of results per page.
+        ///
         /// Specifies the max number of posts to be fetched.
-        /// This is server specific but in regular Lemmy install the default value is 10 and max is 50.
-        public let limit: Int?
+        ///
+        /// - Note: This is server specific but in regular Lemmy install the default value is 10 and max is 50.
+        public let limit: Int64?
 
         /// The community identifier to fetch posts from.
         public let community_id: CommunityId?
@@ -41,8 +45,8 @@ public struct GetPosts: LemmyApiEndpoint {
         public init(
             type_: ListingType? = nil,
             sort: SortType? = nil,
-            page: Int? = nil,
-            limit: Int? = nil,
+            page: Int64? = nil,
+            limit: Int64? = nil,
             community_id: CommunityId? = nil,
             community_name: String? = nil,
             saved_only: Bool? = nil,
